% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.r
\name{data_prepare}
\alias{data_prepare}
\title{A function for preparing the data.frame before the analysis}
\usage{
data_prepare(
  A,
  var_y = NULL,
  vars_cont = NULL,
  vars_disc = NULL,
  vars_int = NULL,
  var_id = NULL
)
}
\arguments{
\item{A}{The data.frame with the variables to test.}

\item{var_y}{The variable for the disease yes/not.}

\item{vars_cont}{The names of the variables with continuous values.}

\item{vars_disc}{The names of the variables with categorical values.}

\item{vars_int}{The names of the variables with integer (ordered) values.}

\item{var_id}{The name of the variable for the unique identifier per row.}
}
\value{
A list with the following entries.
\describe{
  \item{A}{The data.frame from the dataset after checking and updating.}
  \item{var_disc_from_cont}{The names of the discretized variables from continuous ones (not implemented).}
  \item{vars_disc}{The vectors of names received from the input parameters.}
  \item{vars_cont}{The vectors of names received from the input parameters.}
  \item{var_y}{The same name received from the input parameters.}
}
}
\description{
This function takes as input a data.frame and the names of the 
variables in order check the variables and their r types.
}
\examples{

data(DebTrivedi)
A         <- DebTrivedi[,c("age","ofp","gender","region","health")]
A$id      <- 1:nrow(A)
A$hospbin <- DebTrivedi$hosp>0
vars_cont <- c("age","ofp")
vars_disc <- c("gender","region","health")
var_id    <- "id"
var_y     <- "hospbin"
fp = data_prepare(A,var_y,vars_cont,vars_disc,var_id)
print(head(fp$A))

}
