% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DebTrivedi.r
\docType{data}
\name{DebTrivedi}
\alias{DebTrivedi}
\title{Dataset of 4406 individuals aged 66 and over with 19 variables}
\format{
An object of class \code{"data.frame"}.
}
\source{
JSTAT (http://www.jstatsoft.org/v27/i08/paper)
}
\usage{
data(DebTrivedi)
}
\description{
Deb and Trivedi (1997) dataset of 4406 individuals, aged 66 and over, who are covered
by Medicare, a public insurance program. Originally prepared for an R package accompanying 
Kleiber and Zeileis (2008) paper and also available as with Zeileis (2006) paper.
}
\examples{
data(DebTrivedi)
head(DebTrivedi)

}
\references{
A. Zeileis et al. (2008) Journal of Statistical Software 27(8):1-25. \cr
Deb, P. and Trivedi, P. (1997). Demand for medical care by the elderly: 
A finite mixture approach. Journal of Applied Econometrics, 12:313-36.
}
\keyword{datasets}
