% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_compute.r
\name{rep_compute}
\alias{rep_compute}
\title{A function for generating the tables for the report from a data.frame}
\usage{
rep_compute(A, var_y, vars_x, vars_cont, vars_disc, vars_int, var_id)
}
\arguments{
\item{A}{The data.frame for the analysis.}

\item{var_y}{The variable for the disease yes/not.}

\item{vars_x}{The variables for the table with glm for or and rr.}

\item{vars_cont}{The name of the variables with continuous values.}

\item{vars_disc}{The name of the variables with categorical values.}

\item{vars_int}{The name of the variables with integer (ordered) values.}

\item{var_id}{The name of the variable for the unique identifier per row.}
}
\value{
A list with the following entries.
\describe{
  \item{Anew}{The new matrix A after pre-traitment with data_prepare().}
  \item{desc_all}{The result from tab_contents().}
  \item{desc_cont}{The result from tab_desc_cont().}
  \item{desc_disc}{The result from tab_desc_disc().}
  \item{desc_biv}{The result from tab_desc2class_cont().}
  \item{test_tt}{The result from tab_tt2classes_cont().}
  \item{test_anova}{The result from tab_ttanova_cont().}
  \item{test_chi2}{The result from tab_chi2all().}
  \item{or}{The result from tab_all2x2() and stat_oddsratio().}
  \item{rr}{The result from tab_all2x2() and stat_relativerisk().}
  \item{gg}{The result from tab_glmorr().}
  \item{fv}{The result from viz_all2x2().}
  \item{args}{The variables at the call of the function.}
}
}
\description{
This function returns tables from a variable for disease 
which is binary and a set of variables which are categorical for exposures.
The descriptive statistics and statistical tests are computed and aggregated 
in several data.frames.
}
\examples{
\dontrun{
data(DebTrivedi)
A=DebTrivedi
A$id=1:nrow(A)
A$hospbin  = as.integer(A$hosp>0)
var_id    = "id"
vars_cont = c("age","ofp","ofnp","opp","opnp","emer","numchron","hosp","school")
vars_disc = c("health","adldiff","region","black","gender","married","employed",
              "privins","medicaid")
vars_int  = NULL 
var_y     = "hospbin" #binary 0/1
label_y   = "hospibin"
A0 = A
fp = data_prepare(A,var_y,vars_cont,vars_disc,var_id)#,discretize_=TRUE)
A  = fp$A
A$age_3cl = as.character(1*(A$age<7.1)+2*(A$age>=7.1&A$age<7.7)+3*(A$age>=7.7))
A$age_3cl[A$age_3cl=="1"]="[6.6, 7.1)"
A$age_3cl[A$age_3cl=="2"]="[7.1, 7.7)"
A$age_3cl[A$age_3cl=="3"]="[7.7,10.9]"
vars_disc = c(vars_disc,"age_3cl")
vars_x =c(vars_disc,fp$var_disc_from_cont)[c(1:9,10)] #only discrete variables relevant
au = dataepi::rep_compute(A, var_y, vars_x, vars_cont, vars_disc, vars_int, var_id)
}

}
