% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_oddsratio.r
\name{stat_oddsratio}
\alias{stat_oddsratio}
\title{A function for computing the odds ratio from a table for an exposure}
\usage{
stat_oddsratio(X)
}
\arguments{
\item{X}{The object of type table of size 2x2.}
}
\value{
A list with the following entries.
\describe{
  \item{stat}{The odds ratio from the table X.}
  \item{SE}{The standard-deviation of the odds ratio from the table X.}
  \item{I.95left}{The left part of the confidence interval at 0.95\%.}
  \item{I.95right}{The right part of the confidence interval at 0.95\%.}
  \item{name}{The name of the statistics, "OR".}
  \item{warning}{A boolean value for or for not having a=0 or b=0 or c=0 or d=0.}
  \item{a}{The value a from the 2x2 input table.}
  \item{b}{The value b from the 2x2 input table.}
  \item{c}{The value c from the 2x2 input table.}
  \item{d}{The value d from the 2x2 input table.}
  \item{table2x2}{The table from the input parameter.}
}
}
\description{
This function computes the odds ratio of a binary variable
from the crosstable or contingency table between a disease and an exposure.
}
\examples{
X=matrix(c(2534,459,487,142),ncol=2,byrow=TRUE)
X=as.table(X)
colnames(X)<-c("0","1")
rownames(X)<-c("0","1")
print(X)
resu_=stat_oddsratio(X)
cat(resu_$name,"=",round(resu_$stat,2),
paste("(",round(resu_$stdstat,2),")",sep=""),
"\n")

}
