% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_all2x2.r
\name{tab_all2x2}
\alias{tab_all2x2}
\title{A function for generating a table with statistics such as odds ratios}
\usage{
tab_all2x2(A, vars_disc, var_y, stat_f = NULL)
}
\arguments{
\item{A}{The data.frame with the data.}

\item{vars_disc}{The name of the variables with categorical values.}

\item{var_y}{The name of the categorical variable.}

\item{stat_f}{A function such as stat_oddsratio,stat_relativerisk,...}
}
\value{
A list with following entries.
\describe{
  \item{tabstat}{A data.frame with eight columns: the variable name, the 
                 modality or level name, the contents of the 2x2 contingency 
                 table with A for a, B for b, C for c and D for D, and the 
                 corresponding statistics from the function stat_... in the 
                 list of parameters, for instance OR and SE_OR for the odds 
                 ratios and their standard-deviations.}
  \item{binarized_1_s}{The binarized variable for or not the variable equal to 
                      modality, in a list of lists.}
  \item{binarized_0_s}{The binarized variable for or not the variable not equal 
                      to modality, in a list of lists.}
  \item{tables2x2}{The contingency tables for all the variables and all the 
                      modalities in the format of a list a lists, from the 
                      function stat_f in the parameters.}
}
}
\description{
This function returns the whole table with the statistics 
odd ratio, risk ratio of any other defined by the user 
from a set of binary variables (exposures) vs one binary variable (disease).
}
\examples{
library(dataepi)
data(DebTrivedi)
A         <- DebTrivedi
A$hospbin <- as.integer(A$hosp>0)
vars_disc <- c("health","gender","region")
var_y     <- "hospbin"
resu_or   <- dataepi::tab_all2x2(A,vars_disc,var_y,dataepi::stat_oddsratio)
print(resu_or$tabstat)

}
