% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_chi2all.r
\name{tab_chi2all}
\alias{tab_chi2all}
\title{A function for computing the chi2 test from several variables of a data.frame}
\usage{
tab_chi2all(A, vars_disc, pvalue_seuil_ = 0.015)
}
\arguments{
\item{A}{The data.frame with the variables to test.}

\item{vars_disc}{The vector with the names of the variables to test.}

\item{pvalue_seuil_}{The maximal p-value for keeping the pairs of variables.}
}
\value{
A list with three entries
\describe{
  \item{tabchi2}{The data.frame with the chi2 tests by pairs of variables 
  with the following columns}
     \describe{
     \item{row}{The variable from the rows of the contingency table.}
     \item{col}{The variable from the cols of the contingency table.}
     \item{nbr}{The number of modalities of the first variable.}
     \item{nbc}{The number of modalities of the second variable.}
     \item{chi2}{The statistics as computed from the chi2 test.}
     \item{df}{The number of free parameters in the chi2 test.}
     \item{p.val}{The p-value from the chi2 test.}
     \item{mnij}{The minimum coun in the cells of the table.}
     \item{p.val.e}{The p-value from the exact Fisher test.}
     \item{pow}{The power (if available) from the chi2 test.}
     \item{nb}{The total number of counts in the table.}
  }
  \item{pairs_no_pchi2}{A data.frame with by rows the pairs of variables with 
                        no chi2 test available because of their corresponding 
                        contingency table.}
  \item{pairs_large_pchi2}{A data.frame with by rows the pairs of variables 
                          with no chi2 test available because their p-value is 
                          larger than the threshold.}
}
}
\description{
This function take as input a data.frame and the names of the categorical 
variables in order to compute the tests and aggregates them in a data.frame.
The resulting data.frame contains the chi-square tests for each variable in 
vars_disc minus the last which appears in the second column. There are two 
loops: k in (1;p-1) while l in (k+1;p) in order to compute the upper part 
to the diagonal.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_disc <- c("health","gender","region")
resu_    <- tab_chi2all(A,vars_disc,0.05)
print(head(resu_$tabchi2))

}
