% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_chi2oneall.r
\name{tab_chi2oneall}
\alias{tab_chi2oneall}
\title{A function for computing the chi2 test from one against several variables}
\usage{
tab_chi2oneall(A, vars_disc, var_y)
}
\arguments{
\item{A}{The data.frame with the variables to test.}

\item{vars_disc}{The vector with the string names of the variables to test.}

\item{var_y}{The string name of one variable.}
}
\value{
A data.frame with with the chi2 tests by pairs of variables 
  with the following columns
  \describe{
  \item{row}{The variables from the input vector of variable names vars_disc.}
  \item{col}{The variable with its name in the input variable var_y.}
  \item{nbr}{The number of modalities of the first variable.}
  \item{nbc}{The number of modalities of the second variable.}
  \item{chi2}{The statistics as computed from the chi2 test.}
  \item{df}{The number of free parameters in the chi2 test.}
  \item{p.val}{The p-value from the chi2 test.}
  \item{mnij}{The minimum coun in the cells of the table.}
  \item{p.val.e}{The p-value from the exact Fisher test.}
  \item{pow}{The power (if available) from the chi2 test.}
  \item{nb}{The total number of counts in the table.}
 }
}
\description{
This function take as input a data.frame and the names of the categorical 
variables plus one additional variable in order to compute the tests 
and aggregates them in a data.frame, without filtering.
The tests are computed with a loop on the whole set in vars_disc.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_disc <- c("health","gender","region")
var_y     <- "hosp"
resu_    <- tab_chi2oneall(A,vars_disc,var_y)
print(head(resu_))
print(head(resu_[resu_$p.val<=0.05,]))

}
