% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_contents.r
\name{tab_contents}
\alias{tab_contents}
\title{A function for generating a very simple description of the variables 
in a data.frame}
\usage{
tab_contents(D)
}
\arguments{
\item{D}{The data.frame with the variables to describe.}
}
\value{
A data.frame with each row for a variable from the input data.frame and 
        with the following columns.
\describe{
  \item{variable}{The variable name from the column names of D.}
  \item{r_class}{The type of the variable from the R langage.}
  \item{nblevels}{The total number of unique observations.}
  \item{nbobs}{The total number of non missing observations.}
}
Warning: with class factor, it may exist empty levels not counted. The 
function may be considered only after the function data_prepare().
}
\description{
This function show the list of variables 
with the name of their classes of variable (numerical,integer,...) in r
and the number of unique values for each variable.
}
\examples{
data(DebTrivedi)
A    <- DebTrivedi
A$id <- 1:nrow(A)
resu_<-tab_contents(A)
print(head(resu_))

}
