% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc2class_cont.r
\name{tab_desc2class_cont}
\alias{tab_desc2class_cont}
\title{A function for generating a table with statistics of continuous variables}
\usage{
tab_desc2class_cont(A, vars_cont, var_y, nbdigits = 2)
}
\arguments{
\item{A}{The data.frame with the data.}

\item{vars_cont}{The names of the continuous variables.}

\item{var_y}{The name of the categorical variable.}

\item{nbdigits}{The number of decimals to keep.}
}
\value{
A data.frame with each row for a continuous variable and with the 
following columns, where <modality> is one of the modalities of the variable 
whose name in written in var_y.
\describe{
  \item{MEAN_<modality>}{The mean of the continuous variable for the modality.}
  \item{STD_<modality>}{The standard-deviation of the variable for the modality.}
  \item{MD_<modality>}{The median of the continuous variable for the modality.}
  \item{MIN_<modality>}{The minimum of the continuous variable for the modality.}
  \item{MAX_<modality>}{The maximum of the continuous variable for the modality.}
  \item{Nnotna__<modality>}{The number of non missing observation for the modality.}
}
}
\description{
This function returns the whole table with the statistics 
number of observations, mean, standard-error, median, min, max
from a set of continuous variables versus one categorical variable 
with two (or eventually more modalities).
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_cont <- c("age","ofp")
resu_    <- tab_desc2class_cont(A,vars_cont,"gender")
print(resu_)

}
