% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc_cont.r
\name{tab_desc_cont}
\alias{tab_desc_cont}
\title{A function for describing the data.frame for the continuous variables}
\usage{
tab_desc_cont(A, vars_cont, nbdigits = 2)
}
\arguments{
\item{A}{The data.frame with the variables to test.}

\item{vars_cont}{The names of the variables with continuous values.}

\item{nbdigits}{The number of decimals to keep.}
}
\value{
A data.frame with each row for a continuous variable and with the 
following columns.
\describe{
  \item{var}{The name of a variable from the vector of names vars_cont.}
  \item{median}{The median of the variable.}
  \item{mean}{The mean of the variable.}
  \item{sd}{The standard-deviation of the variable.}
  \item{min}{The minimum of the variable.}
  \item{max}{The maximum of the variable.}
  \item{nb_na}{The number of missing values of the variable.}
  \item{nb}{The number of non missing values of the variable.}
}
}
\description{
This function take as input a data.frame and the names of the continuous  
variables in order to print in a table the names, plus the median, mean,
standard-deviation, minimum, maximum, number of NA, and the number of not NA.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_cont <- c("age","ofp")
resu_    <- tab_desc_cont(A,vars_cont)
print(resu_) 

}
