% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_desc_disc.r
\name{tab_desc_disc}
\alias{tab_desc_disc}
\title{A function for describing the data.frame for the categorical variables}
\usage{
tab_desc_disc(A, vars_disc, nbdigits = 2)
}
\arguments{
\item{A}{The data.frame with the variables to test.}

\item{vars_disc}{The names of the variables with categorical values.}

\item{nbdigits}{The number of decimals to keep.}
}
\value{
A data.frame with each row for a categorical variable and with the 
following columns.
\describe{
  \item{var}{The name of a variable from the vector of names vars_disc.}
  \item{nb_na}{The number of missing values.}
  \item{nblevel}{The total number of unique observations.}
  \item{nbperlevel}{The numbers of observations by modality.}
  \item{propperlevel}{The proportions of observations by modality.}
  \item{namelevel}{The corresponding modality names.}
}
}
\description{
This function take as input a data.frame and the names of the categorical 
variables in order to print in a table the names, frequencies per modalities,
percentable per modalities, and the number of levels and number of NA values.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_disc <- c("health","gender","region")
resu_    <- tab_desc_disc(A,vars_disc)
print(resu_) 

}
