% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_glmorr.r
\name{tab_glmorr}
\alias{tab_glmorr}
\title{A function for computing the OR from the logistic regression}
\usage{
tab_glmorr(A, vars_x, var_y)
}
\arguments{
\item{A}{The data.frame for the analysis.}

\item{vars_x}{The variables for the table with glm for or and rr.}

\item{var_y}{The variable for the disease yes/not.}
}
\value{
A list with following entries.
\describe{
  \item{frm}{The formula for the logistic regression for computing the 
             odds ratio on a full model before selecting the variables.}
  \item{tabl_coeff_full}{The resulting table of coefficients regression 
                         with all the variables.}
  \item{tabl_coeff_small}{The resulting table of coefficients regression 
                         with the variable kept after a selection by AIC.}
  \item{allcoeffs}{The two tables of coefficients side to side for 
                   comparison purpose.}
  \item{allors}{The odds ratios from the two tables of coefficients side to 
                 side in allcoeffs.}
  \item{fit_full}{The r object from the glm regression with all variables.}
  \item{fit_small}{The r object from the glm regression with selected variables.}
  \item{yX}{The dataframe restricted to the variable in vars_x and var_y.}
}
}
\description{
This function take the data.frame with all the variable and 
construct a new matrix with the odds ratio from a logistic regression,
from the full model and from the reduced model after a selection by AIC.
}
\examples{
\dontrun{
data(DebTrivedi)
A=DebTrivedi
A$hospbin = as.integer(A$hosp>0)
vars_x = c("health","region","gender","married","employed")
var_y  = "hospbin"
gg     = dataepi::tab_glmorr(A,vars_x,var_y) 
print(gg$allors)
}

}
