% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_tt2classes_cont.r
\name{tab_tt2classes_cont}
\alias{tab_tt2classes_cont}
\title{A function for generating a table with the t-test of diverse variables}
\usage{
tab_tt2classes_cont(A, vars_cont, var_y)
}
\arguments{
\item{A}{The data.frame with the data.}

\item{vars_cont}{The names of the continuous variables.}

\item{var_y}{The name of the categorical variable.}
}
\value{
A data.frame with each row for a continuous variable and with the 
following columns.
\describe{
  \item{var1}{The name of the continuous variable from group 1.}
  \item{median1}{The median of the continuous variable from group 1.}
  \item{mean1}{The mean of the continuous variable from group 1.}
  \item{sd1}{The standard-deviation of the continuous variable from group 1.}
  \item{nb1}{The sample size of group 1.}
  \item{var2}{The name of the continuous variable from group 2.}
  \item{median2}{The median of the continuous variable from group 2.}
  \item{mean2}{The mean of the continuous variable from group 2.}
  \item{sd2}{The standard-deviation of the continuous variable from group 2.}
  \item{nb2}{The sample size of group 2.}
  \item{T_t.test (2cl)}{The statistics computed for the t-Student test.}
  \item{P_t.test (2cl)}{The p-value computed for the t-Student test.}
  \item{P<0.05 Power_t.t}{The power computed for the t-Student test.}
}
}
\description{
This function returns the whole table with the t-tests 
from a set of continuous variables versus one categorical variable 
with two (or eventually more modalities).
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_cont <- c("age","ofp")
resu_    <- tab_tt2classes_cont(A,vars_cont,"gender")
print(resu_)

}
