% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ttanova_cont.r
\name{tab_ttanova_cont}
\alias{tab_ttanova_cont}
\title{A function for generating a table with the anova of diverse variables}
\usage{
tab_ttanova_cont(A, vars_cont, vars_disc)
}
\arguments{
\item{A}{The data.frame with the data.}

\item{vars_cont}{The names of the continuous variables.}

\item{vars_disc}{The names of the categorical variables.}
}
\value{
A list with the following entries.
\describe{
  \item{tabstat}{
A data.frame with each row for a continuous variable and a 
categorical variable (binary or poly) with the following columns.
\describe{
  \item{var_cont}{The name of the continuous variable.}
  \item{var_disc}{The name of the discrete variable.}
  \item{n1}{The size of groupe 1.}
  \item{n2}{The size of groupe 2.}
  \item{...}{...}
  \item{ng}{The size of groupe g (if it exists for the variable in var_disc).}
  \item{norm1}{The p-value of the normality test from group 1.}
  \item{norm2}{The p-value of the normality test from group 2.}
  \item{...}{...}
  \item{normg}{The p-value of the normality test from group g (if exists).}
  \item{test12_vr}{The p-value of the equality test of the variances for two groups.}
  \item{test12_eq}{The p-value of the equality t-test of the means for two groups.}
  \item{test12_gt}{The p-value of the equality t-test of the means for two groups with option "greater".}
  \item{test12_ls}{The p-value of the equality t-test of the means for two groups with option "less".}
  \item{test12_wx_eq}{The p-value of the wilkox rank sum (not paired) equality test of the means for two groups.}
  \item{test12_wx_gq}{The p-value of the wilkox rank sum (not paired) equality test of the means for two groups  with option "greater".}
  \item{test12_wx_ls}{The p-value of the wilkox rank sum (not paired) equality test of the means for two groups  with option "less".}
  \item{test_vr}{The p-value of the equality test of the variances for g>2 groups.}
  \item{test_aov}{The p-value of the equality test of the means for g>2 groups.}
  \item{test_aov_check}{The p-value from the normality test of the residual from the anova test for g>2 groups.}
  \item{test_welch}{The p-value from the welch test for g>=2 groups and variances not equal, under normality and variances supposed not equal.}
  \item{test_krusk}{The p-value from the kruskal-wallis rank sum test for g>=2 groups to check the equality in distribution of the means.}
 }
}
\item{pairs_no_panova}{A list of pairs of variables with non available anova test.}
}
}
\description{
This function returns the whole table with the anova tests 
from a set of continuous variables and categorical variables 
with different number of modalities, plus other related tests.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
vars_cont <- c("age","ofp")
vars_disc <- c("gender","region","health")
resu_    <- tab_ttanova_cont(A,vars_cont,vars_disc)
print(resu_)

}
