% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_all2x2.r
\name{viz_all2x2}
\alias{viz_all2x2}
\title{A function for showing the modalities in a 2d view plus a related indicator}
\usage{
viz_all2x2(TABSTAT, g_ = 3, graph_ = "PCA")
}
\arguments{
\item{TABSTAT}{The table from the list of 2x2 tables from the data.}

\item{g_}{The number of groups for the clustering.}

\item{graph_}{The characters string with null value for no projection, with the 
value to show the "PCA", the value "COSTAT" to show the percentages for 
Sick=1, and the value "CASTAT" to show the percentages for the Sick=0.}
}
\value{
A list with the following entries.
\describe{
  \item{tabstat}{A data.frame with in the first left columns the matrix from the tables 
                2x2 aggregated from the function tab_all2x2() after normalization of its pairs 
                of columns, followed by two columns, one for the variable name and one for the 
                modality name.}
  \item{pca}{The r object from a pca of the normalized matrix of counts.}
  \item{kmeans}{The r object from a kmeans of the normalized matrix of counts.}
  \item{ORpca}{An alternative indicator for odds ratios.}
}
}
\description{
This function take the set of 2x2 table for the odds ratio in order to 
construct a new matrix of the whole set of modalities and performs pca.
The obtained projection is visualized and an indicator is also obtained.
}
\examples{
data(DebTrivedi)
A         <- DebTrivedi
A$hospbin <- as.integer(A$hosp>0)
vars_disc <- c("health","gender","region")
var_y     <- "hospbin"
resu_or   <- dataepi::tab_all2x2(A,vars_disc,var_y,dataepi::stat_oddsratio)
resu_fv   <- viz_all2x2(resu_or$tabstat,g_=7,graph_="PCA")
print(resu_fv$tabstat)

}
